requirements.txt
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
drf-yasg==1.21.7
pillow==10.3.0

